pyproject.toml
[project]
name = "py-message-bus"
version = "0.1.0"
description = "Lightweight message bus for modular monolith architecture"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["message-bus", "cqrs", "modular-monolith", "clean-architecture"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]

dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "import-linter>=2.0",
    "pre-commit>=3.0",
]
kafka = [
    "aiokafka>=0.10",
]
zmq = [
    "pyzmq>=25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/message_bus"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = ["import-untyped", "no-untyped-def", "misc"]

# Import Linter - 아키텍처 의존성 규칙
[tool.importlinter]
root_packages = ["message_bus"]
include_external_packages = true

[[tool.importlinter.contracts]]
name = "ports should not depend on implementations"
type = "forbidden"
source_modules = ["message_bus.ports"]
forbidden_modules = ["message_bus.local", "message_bus.zmq_bus"]
